Replace XCTest with Quick Spec for LocationUpdater
